Split tests/fleet/test_dispatch_failure_semantics.py along concern boundaries
Extract 5 shared helpers to tests/fleet/_helpers.py (_run,
_read_dispatch_record, _make_no_sentinel, _make_completed_dirty,
_make_completed_clean) eliminating existing duplication with _helpers.py.

Create 6 new companion files from the original 1,247-line file:
- test_dispatch_envelope_fields.py: TestElapsedSecondsEnvelopeField,
  TestDispatchStatusEnvelopeField (5 tests)
- test_dispatch_stderr_forwarding.py: TestStderrEnvelopeForwarding (10 tests)
- test_dispatch_ingredient_validation.py: TestMissingRequiredIngredient (8 tests)
- test_dispatch_recipe_kind_gate.py: TestRecipeKindDispatchGate (2 tests)
- test_dispatch_crash_diagnostics.py: TestCrashPathDiagnosticPersistence (3 tests)
- test_dispatch_labels_cleaned.py: TestLabelsCleanedFieldPersistence (4 tests)

Trim original to 4 core failure path classes (12 tests):
TestTimeoutPath, TestNoSentinelPath, TestCompletedDirtyPath,
TestCompletedCleanPath (~350 lines).

Update test filter cascade (_test_filter.py) to route
test_dispatch_ingredient_validation.py and test_dispatch_recipe_kind_gate.py
to recipe cascade, and test_dispatch_crash_diagnostics.py to server cascade.

Add _FLEET_SERVER_FILE_LEVEL_ENTRIES guard and corresponding test in
test_test_filter_cascade.py for the server cascade fleet entry.

Update tests/fleet/CLAUDE.md with 7 entries replacing the single
original entry.
