+++ tests/fleet/test_dispatch_failure_semantics.py
@@ -1,100 +1,21 @@
-"""Group F: Timeout + No-Result-Block failure semantics for fleet dispatch."""
+"""Group F: Core failure path semantics — timeout, no-sentinel, completed-dirty, completed-clean."""
 
 from __future__ import annotations
-
-import json
-from pathlib import Path
-from unittest.mock import AsyncMock
 
 import pytest
 
 from autoskillit.core.types import CliSubtype
-from autoskillit.fleet.result_parser import L3ParseResult
 from tests.fakes import InMemoryHeadlessExecutor
-from tests.fleet._helpers import _setup_dispatch
+from tests.fleet._helpers import (
+    _make_completed_clean,
+    _make_completed_dirty,
+    _make_no_sentinel,
+    _read_dispatch_record,
+    _run,
+    _setup_dispatch,
+)
 
 pytestmark = [pytest.mark.layer("fleet"), pytest.mark.small, pytest.mark.feature("fleet")]
-
-
-def _simple_prompt_builder(**kwargs) -> str:
-    return f"prompt-for-{kwargs.get('recipe', 'unknown')}"
-
-
-async def _no_sleep_quota_checker(config, **kwargs) -> dict:
-    return {
-        "should_sleep": False,
-        "sleep_seconds": 0,
-        "utilization": None,
-        "resets_at": None,
-        "window_name": None,
-    }
-
-
-async def _noop_quota_refresher(config, **kwargs) -> None:
-    pass
-
-
-async def _run(
-    tool_ctx, recipe: str = "test-recipe", ingredients: dict[str, str] | None = None
-) -> dict:
-    from autoskillit.fleet._api import execute_dispatch
-
-    result = await execute_dispatch(
-        tool_ctx=tool_ctx,
-        recipe=recipe,
-        task="t",
-        ingredients=ingredients,
-        dispatch_name=None,
-        timeout_sec=None,
-        prompt_builder=_simple_prompt_builder,
-        quota_checker=_no_sleep_quota_checker,
-        quota_refresher=_noop_quota_refresher,
-    )
-    return json.loads(result.outcome.to_envelope())
-
-
-def _read_dispatch_record(tool_ctx) -> dict:
-    """Read the single dispatch record written to the state file."""
-    state_files = list((tool_ctx.temp_dir / "dispatches").glob("*.json"))
-    assert len(state_files) == 1, f"Expected 1 state file, found {len(state_files)}"
-    state = json.loads(state_files[0].read_text())
-    return state["dispatches"][0]
-
-
-def _make_no_sentinel():
-
-    return L3ParseResult(
-        outcome="no_sentinel",
-        payload=None,
-        raw_body=None,
-        parse_error=None,
-        source="stdout",
-    )
-
-
-def _make_completed_dirty():
-
-    return L3ParseResult(
-        outcome="completed_dirty",
-        payload=None,
-        raw_body="garbled",
-        parse_error="json decode error",
-        source="stdout",
-    )
-
-
-def _make_completed_clean(success: bool, reason: str = ""):
-
-    payload: dict = {"success": success}
-    if reason:
-        payload["reason"] = reason
-    return L3ParseResult(
-        outcome="completed_clean",
-        payload=payload,
-        raw_body=None,
-        parse_error=None,
-        source="stdout",
-    )
 
 
 class TestTimeoutPath:
@@ -339,126 +260,6 @@ class TestCompletedDirtyPath:
         assert record["reason"] == "fleet_l3_parse_failed"
 
 
-class TestElapsedSecondsEnvelopeField:
-    @pytest.mark.anyio
-    async def test_dispatch_completed_envelope_contains_elapsed_seconds(self):
-        """DispatchCompleted.to_envelope() includes elapsed_seconds field.
-
-        Regression test: elapsed_seconds is computed in _run_dispatch() but never
-        added to DispatchCompleted, so the L3 agent has no timing source.
-        """
-        from autoskillit.fleet import DispatchCompleted, DispatchStatus
-
-        completed = DispatchCompleted(
-            success=True,
-            dispatch_status=DispatchStatus.SUCCESS,
-            dispatch_id="test-dispatch-id",
-            dispatched_session_id="test-session-id",
-            reason="",
-            token_usage={"input": 100, "output": 50, "cache_read": 10, "cache_creation": 5},
-            elapsed_seconds=42.5,
-        )
-        result = json.loads(completed.to_envelope())
-        assert "elapsed_seconds" in result, "to_envelope() must include elapsed_seconds"
-        assert result["elapsed_seconds"] == 42.5
-
-
-class TestDispatchStatusEnvelopeField:
-    @pytest.mark.anyio
-    async def test_envelope_includes_dispatch_status_on_success(self, tool_ctx, monkeypatch):
-        """Envelope from _run_dispatch includes dispatch_status matching state-file status."""
-        _setup_dispatch(tool_ctx, monkeypatch)
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_completed_clean(success=True),
-        )
-
-        result = await _run(tool_ctx)
-        assert "dispatch_status" in result
-        assert result["dispatch_status"] == "success"
-
-    @pytest.mark.anyio
-    async def test_envelope_includes_dispatch_status_on_failure(self, tool_ctx, monkeypatch):
-        """Envelope includes dispatch_status='failure' when outcome is completed_dirty."""
-        _setup_dispatch(tool_ctx, monkeypatch)
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_completed_dirty(),
-        )
-
-        result = await _run(tool_ctx)
-        assert "dispatch_status" in result
-        assert result["dispatch_status"] == "failure"
-
-    @pytest.mark.anyio
-    async def test_envelope_includes_dispatch_status_on_no_sentinel(self, tool_ctx, monkeypatch):
-        """Envelope includes dispatch_status='failure' for no_sentinel without session signal."""
-        _setup_dispatch(tool_ctx, monkeypatch)
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_no_sentinel(),
-        )
-
-        result = await _run(tool_ctx)
-        assert "dispatch_status" in result
-        assert result["dispatch_status"] == "failure"
-
-    @pytest.mark.anyio
-    async def test_envelope_includes_dispatch_status_on_no_sentinel_resumable(
-        self, tool_ctx, monkeypatch
-    ):
-        """no_sentinel + session_id + lifespan_started + sidecar → dispatch_status='resumable'."""
-        import dataclasses
-
-        from autoskillit.core import DispatchIdentity
-        from tests.fakes import _DEFAULT_SKILL_RESULT, InMemoryHeadlessExecutor
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-
-        fixed_dispatch_id = "aaaabbbb-cccc-dddd-eeee-ffffffffffff"
-        _fixed_identity = DispatchIdentity.from_dispatch_id(fixed_dispatch_id)
-
-        class _FixedDispatchIdentity:
-            @classmethod
-            def fresh(cls) -> DispatchIdentity:
-                return _fixed_identity
-
-        monkeypatch.setattr("autoskillit.fleet.state.DispatchIdentity", _FixedDispatchIdentity)
-
-        from autoskillit.fleet.sidecar import sidecar_path
-
-        sidecar_file = sidecar_path(fixed_dispatch_id, tool_ctx.project_dir)
-        sidecar_file.parent.mkdir(parents=True, exist_ok=True)
-        sidecar_file.touch()
-
-        resumable_result = dataclasses.replace(
-            _DEFAULT_SKILL_RESULT,
-            session_id="sess-resumable-abc",
-            lifespan_started=True,
-        )
-
-        class _SpawningExecutor(InMemoryHeadlessExecutor):
-            """Calls on_spawn with a fake PID to drive PENDING → RUNNING before returning."""
-
-            async def dispatch_food_truck(self, orchestrator_prompt, cwd, *, on_spawn=None, **kw):
-                if on_spawn is not None:
-                    on_spawn(12345, 1000)
-                return await super().dispatch_food_truck(
-                    orchestrator_prompt, cwd, on_spawn=on_spawn, **kw
-                )
-
-        tool_ctx.executor = _SpawningExecutor(default_result=resumable_result)
-
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_no_sentinel(),
-        )
-
-        result = await _run(tool_ctx)
-        assert "dispatch_status" in result
-        assert result["dispatch_status"] == "resumable"
-
-
 class TestCompletedCleanPath:
     @pytest.mark.anyio
     async def test_completed_clean_success_writes_empty_reason(self, tool_ctx, monkeypatch):
@@ -487,760 +288,3 @@ class TestCompletedCleanPath:
 
         record = _read_dispatch_record(tool_ctx)
         assert record["reason"] == "my-failure-reason"
-
-
-# ---------------------------------------------------------------------------
-# Group: Missing required ingredient validation
-# ---------------------------------------------------------------------------
-
-
-def _setup_dispatch_with_ingredients(tool_ctx, monkeypatch, ingredients: dict):
-    """Wire tool_ctx with a recipe that has specific ingredients."""
-    from autoskillit.fleet import FleetSemaphore
-    from autoskillit.recipe.schema import Recipe, RecipeIngredient, RecipeKind
-    from tests.fakes import InMemoryHeadlessExecutor, InMemoryRecipeRepository
-
-    tool_ctx.fleet_lock = FleetSemaphore(max_concurrent=1)
-    repo = InMemoryRecipeRepository()
-    recipe_info = _make_recipe_info("test-recipe")
-    repo.add_recipe("test-recipe", recipe_info)
-    repo.add_full_recipe(
-        recipe_info.path,
-        Recipe(
-            name="test-recipe",
-            description="test",
-            kind=RecipeKind.STANDARD,
-            ingredients={
-                k: RecipeIngredient(description=f"desc-{k}", **v) for k, v in ingredients.items()
-            },
-        ),
-    )
-    tool_ctx.recipes = repo
-    tool_ctx.executor = InMemoryHeadlessExecutor()
-
-
-def _make_recipe_info(name: str = "test-recipe"):
-    from tests.fleet._helpers import _make_recipe_info as _base
-
-    return _base(name)
-
-
-class TestMissingRequiredIngredient:
-    @pytest.mark.anyio
-    async def test_dispatch_rejects_missing_required_ingredient(self, tool_ctx, monkeypatch):
-        """Required ingredient with no default → FLEET_MISSING_INGREDIENT."""
-        _setup_dispatch_with_ingredients(
-            tool_ctx, monkeypatch, {"api_key": {"required": True, "default": None}}
-        )
-
-        result = await _run(tool_ctx, ingredients={})
-        assert result["success"] is False
-        assert result["error"] == "fleet_missing_ingredient"
-
-    @pytest.mark.anyio
-    async def test_dispatch_allows_required_ingredient_when_supplied(self, tool_ctx, monkeypatch):
-        """A required ingredient that IS supplied passes validation."""
-        _setup_dispatch_with_ingredients(
-            tool_ctx, monkeypatch, {"api_key": {"required": True, "default": None}}
-        )
-
-        result = await _run(tool_ctx, ingredients={"api_key": "secret"})
-        assert result.get("error") != "fleet_missing_ingredient"
-
-    @pytest.mark.anyio
-    async def test_dispatch_allows_required_ingredient_with_default(self, tool_ctx, monkeypatch):
-        """A required ingredient with a non-None default passes even when not supplied."""
-        _setup_dispatch_with_ingredients(
-            tool_ctx, monkeypatch, {"api_key": {"required": True, "default": "fallback"}}
-        )
-
-        result = await _run(tool_ctx, ingredients={})
-        assert result.get("error") != "fleet_missing_ingredient"
-
-    @pytest.mark.anyio
-    async def test_dispatch_lists_all_missing_required_ingredients(self, tool_ctx, monkeypatch):
-        """When multiple required ingredients are missing, all are listed."""
-        _setup_dispatch_with_ingredients(
-            tool_ctx,
-            monkeypatch,
-            {
-                "key_a": {"required": True, "default": None},
-                "key_b": {"required": True, "default": None},
-            },
-        )
-
-        result = await _run(tool_ctx, ingredients={})
-        assert result["success"] is False
-        assert result["error"] == "fleet_missing_ingredient"
-        assert "key_a" in result["user_visible_message"]
-        assert "key_b" in result["user_visible_message"]
-
-    @pytest.mark.anyio
-    async def test_dispatch_ignores_optional_missing_ingredients(self, tool_ctx, monkeypatch):
-        """Optional ingredients (required=False) don't trigger missing-ingredient errors."""
-        _setup_dispatch_with_ingredients(
-            tool_ctx, monkeypatch, {"optional_key": {"required": False, "default": None}}
-        )
-
-        result = await _run(tool_ctx, ingredients={})
-        assert result.get("error") != "fleet_missing_ingredient"
-
-    @pytest.mark.anyio
-    async def test_task_auto_injected_from_top_level_param(self, tool_ctx, monkeypatch):
-        """top-level task param auto-injects into effective_ingredients when recipe declares it."""
-        _setup_dispatch_with_ingredients(
-            tool_ctx, monkeypatch, {"task": {"required": True, "default": None}}
-        )
-
-        result = await _run(tool_ctx, ingredients={})
-        assert result.get("error") != "fleet_missing_ingredient"
-
-    @pytest.mark.anyio
-    async def test_explicit_ingredient_task_overrides_top_level(self, tool_ctx, monkeypatch):
-        """Explicit ingredients['task'] takes precedence over top-level task param."""
-        _setup_dispatch_with_ingredients(
-            tool_ctx, monkeypatch, {"task": {"required": True, "default": None}}
-        )
-
-        captured = {}
-
-        def _capture_prompt_builder(**kwargs):
-            captured.update(kwargs)
-            return "prompt"
-
-        from autoskillit.fleet._api import execute_dispatch
-
-        await execute_dispatch(
-            tool_ctx=tool_ctx,
-            recipe="test-recipe",
-            task="top-level-value",
-            ingredients={"task": "override-value"},
-            dispatch_name=None,
-            timeout_sec=None,
-            prompt_builder=_capture_prompt_builder,
-            quota_checker=_no_sleep_quota_checker,
-            quota_refresher=_noop_quota_refresher,
-        )
-
-        assert captured["ingredients"]["task"] == "override-value"
-
-    @pytest.mark.anyio
-    async def test_task_not_injected_when_not_declared_ingredient(self, tool_ctx, monkeypatch):
-        """Top-level task is NOT injected when recipe has no 'task' ingredient key."""
-        _setup_dispatch_with_ingredients(
-            tool_ctx, monkeypatch, {"other_key": {"required": False, "default": "x"}}
-        )
-
-        captured = {}
-
-        def _capture_prompt_builder(**kwargs):
-            captured.update(kwargs)
-            return "prompt"
-
-        from autoskillit.fleet._api import execute_dispatch
-
-        await execute_dispatch(
-            tool_ctx=tool_ctx,
-            recipe="test-recipe",
-            task="some-task",
-            ingredients={},
-            dispatch_name=None,
-            timeout_sec=None,
-            prompt_builder=_capture_prompt_builder,
-            quota_checker=_no_sleep_quota_checker,
-            quota_refresher=_noop_quota_refresher,
-        )
-
-        assert "task" not in captured["ingredients"]
-
-
-# ---------------------------------------------------------------------------
-# Group: Recipe kind dispatch gate
-# ---------------------------------------------------------------------------
-
-
-class TestRecipeKindDispatchGate:
-    """Verify that dispatch gate accepts/rejects by RecipeKind."""
-
-    def _setup_food_truck_recipe(self, tool_ctx):
-        """Wire tool_ctx with a food-truck kind recipe."""
-        from autoskillit.fleet import FleetSemaphore
-        from autoskillit.recipe.schema import Recipe, RecipeKind
-        from tests.fakes import InMemoryHeadlessExecutor, InMemoryRecipeRepository
-
-        tool_ctx.fleet_lock = FleetSemaphore(max_concurrent=1)
-        repo = InMemoryRecipeRepository()
-        recipe_info = _make_recipe_info("test-recipe")
-        repo.add_recipe("test-recipe", recipe_info)
-        repo.add_full_recipe(
-            recipe_info.path,
-            Recipe(
-                name="test-recipe",
-                description="test",
-                kind=RecipeKind.FOOD_TRUCK,
-                ingredients={},
-                requires_packs=[],
-            ),
-        )
-        tool_ctx.recipes = repo
-        tool_ctx.executor = InMemoryHeadlessExecutor()
-
-    def _setup_campaign_recipe(self, tool_ctx):
-        """Wire tool_ctx with a campaign kind recipe."""
-        from autoskillit.fleet import FleetSemaphore
-        from autoskillit.recipe.schema import Recipe, RecipeKind
-        from tests.fakes import InMemoryHeadlessExecutor, InMemoryRecipeRepository
-
-        tool_ctx.fleet_lock = FleetSemaphore(max_concurrent=1)
-        repo = InMemoryRecipeRepository()
-        recipe_info = _make_recipe_info("test-recipe")
-        repo.add_recipe("test-recipe", recipe_info)
-        repo.add_full_recipe(
-            recipe_info.path,
-            Recipe(
-                name="test-recipe",
-                description="test",
-                kind=RecipeKind.CAMPAIGN,
-                ingredients={},
-                requires_packs=[],
-            ),
-        )
-        tool_ctx.recipes = repo
-        tool_ctx.executor = InMemoryHeadlessExecutor()
-
-    @pytest.mark.anyio
-    async def test_food_truck_dispatchable(self, tool_ctx, monkeypatch):
-        """T4: FOOD_TRUCK kind is accepted by the dispatch gate (not rejected)."""
-        self._setup_food_truck_recipe(tool_ctx)
-
-        result = await _run(tool_ctx)
-        # Any error OTHER than fleet_invalid_recipe_kind means the gate passed
-        assert result.get("error") != "fleet_invalid_recipe_kind"
-
-    @pytest.mark.anyio
-    async def test_campaign_kind_still_rejected_by_dispatch(self, tool_ctx, monkeypatch):
-        """T5: CAMPAIGN kind is still rejected by the dispatch gate."""
-        self._setup_campaign_recipe(tool_ctx)
-
-        result = await _run(tool_ctx)
-        assert result["success"] is False
-        assert result["error"] == "fleet_invalid_recipe_kind"
-
-
-class TestStderrEnvelopeForwarding:
-    """Verify that stderr from SkillResult is forwarded in all dispatch envelope shapes."""
-
-    @pytest.mark.anyio
-    async def test_completed_clean_envelope_includes_stderr(self, tool_ctx, monkeypatch):
-        """completed_clean envelope includes stderr field with skill_result.stderr value."""
-        import dataclasses
-
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        tool_ctx.executor = InMemoryHeadlessExecutor(
-            default_result=dataclasses.replace(_DEFAULT_SKILL_RESULT, stderr="some error output")
-        )
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_completed_clean(success=True),
-        )
-
-        result = await _run(tool_ctx)
-        assert result["stderr"] == "some error output"
-
-    @pytest.mark.anyio
-    async def test_completed_dirty_envelope_includes_stderr(self, tool_ctx, monkeypatch):
-        """completed_dirty envelope includes stderr field with skill_result.stderr value."""
-        import dataclasses
-
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        tool_ctx.executor = InMemoryHeadlessExecutor(
-            default_result=dataclasses.replace(_DEFAULT_SKILL_RESULT, stderr="parse failure trace")
-        )
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_completed_dirty(),
-        )
-
-        result = await _run(tool_ctx)
-        assert result["stderr"] == "parse failure trace"
-
-    @pytest.mark.anyio
-    async def test_no_sentinel_envelope_includes_stderr(self, tool_ctx, monkeypatch):
-        """no_sentinel envelope includes stderr field with skill_result.stderr value."""
-        import dataclasses
-
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        tool_ctx.executor = InMemoryHeadlessExecutor(
-            default_result=dataclasses.replace(
-                _DEFAULT_SKILL_RESULT, stderr="missing sentinel trace"
-            )
-        )
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_no_sentinel(),
-        )
-
-        result = await _run(tool_ctx)
-        assert result["stderr"] == "missing sentinel trace"
-
-    @pytest.mark.anyio
-    async def test_no_sentinel_envelope_stderr_truncated_to_envelope_max(
-        self, tool_ctx, monkeypatch
-    ):
-        """no_sentinel envelope stderr is truncated when stderr exceeds ENVELOPE_STDERR_MAX."""
-        import dataclasses
-
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        long_stderr = "x" * 3000
-        tool_ctx.executor = InMemoryHeadlessExecutor(
-            default_result=dataclasses.replace(_DEFAULT_SKILL_RESULT, stderr=long_stderr)
-        )
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_no_sentinel(),
-        )
-
-        result = await _run(tool_ctx)
-        assert len(result["stderr"]) < 3000
-        assert "truncated" in result["stderr"]
-
-    @pytest.mark.anyio
-    async def test_timeout_envelope_includes_stderr_in_details(self, tool_ctx, monkeypatch):
-        """Timeout fleet_error envelope includes stderr in details dict."""
-        import dataclasses
-
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        tool_ctx.executor = InMemoryHeadlessExecutor(
-            default_result=dataclasses.replace(
-                _DEFAULT_SKILL_RESULT, subtype="timeout", stderr="timeout stderr"
-            )
-        )
-
-        result = await _run(tool_ctx)
-        assert result["stderr"] == "timeout stderr"
-
-    @pytest.mark.anyio
-    async def test_timeout_envelope_stderr_truncated_to_envelope_max(self, tool_ctx, monkeypatch):
-        """Timeout envelope stderr is truncated when stderr exceeds ENVELOPE_STDERR_MAX."""
-        import dataclasses
-
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        long_stderr = "x" * 3000
-        tool_ctx.executor = InMemoryHeadlessExecutor(
-            default_result=dataclasses.replace(
-                _DEFAULT_SKILL_RESULT, subtype="timeout", stderr=long_stderr
-            )
-        )
-
-        result = await _run(tool_ctx)
-        assert len(result["stderr"]) < 3000
-        assert "truncated" in result["stderr"]
-
-    @pytest.mark.anyio
-    async def test_envelope_stderr_truncated_to_envelope_max(self, tool_ctx, monkeypatch):
-        """When skill_result.stderr exceeds ENVELOPE_STDERR_MAX, envelope truncates it."""
-        import dataclasses
-
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        long_stderr = "x" * 3000
-        tool_ctx.executor = InMemoryHeadlessExecutor(
-            default_result=dataclasses.replace(
-                _DEFAULT_SKILL_RESULT, stderr=long_stderr, success=False
-            )
-        )
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_completed_clean(success=False),
-        )
-
-        result = await _run(tool_ctx)
-        assert len(result["stderr"]) < 3000
-        assert "truncated" in result["stderr"]
-
-    @pytest.mark.anyio
-    async def test_completed_dirty_envelope_stderr_truncated_to_envelope_max(
-        self, tool_ctx, monkeypatch
-    ):
-        """completed_dirty envelope stderr is truncated when stderr exceeds ENVELOPE_STDERR_MAX."""
-        import dataclasses
-
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        long_stderr = "x" * 3000
-        tool_ctx.executor = InMemoryHeadlessExecutor(
-            default_result=dataclasses.replace(
-                _DEFAULT_SKILL_RESULT, stderr=long_stderr, success=False
-            )
-        )
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_completed_dirty(),
-        )
-
-        result = await _run(tool_ctx)
-        assert len(result["stderr"]) < 3000
-        assert "truncated" in result["stderr"]
-
-    @pytest.mark.anyio
-    async def test_envelope_stderr_empty_when_no_stderr(self, tool_ctx, monkeypatch):
-        """When skill_result.stderr is empty, envelope stderr field is empty string."""
-        _setup_dispatch(tool_ctx, monkeypatch)
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_completed_clean(success=True),
-        )
-
-        result = await _run(tool_ctx)
-        assert result["stderr"] == ""
-
-
-class TestCrashPathDiagnosticPersistence:
-    """When _run_dispatch raises, the exception message must be persisted to campaign state."""
-
-    async def test_crash_persists_exception_message_to_campaign_state(
-        self, tool_ctx, monkeypatch, tmp_path: Path
-    ) -> None:
-        """Inject RuntimeError into _run_dispatch; verify campaign state carries the message."""
-        from autoskillit.fleet import (
-            DispatchRecord,
-            read_state,
-            write_initial_state,
-        )
-        from tests.fakes import InMemoryHeadlessExecutor
-
-        campaign_state_path = tmp_path / "campaign.json"
-        write_initial_state(
-            campaign_state_path,
-            "cmp-crash",
-            "test-campaign",
-            "/m.yaml",
-            [DispatchRecord(name="dispatch-a")],
-        )
-        _setup_dispatch(tool_ctx, monkeypatch)
-        executor = InMemoryHeadlessExecutor()
-
-        # Inject a crash: any dispatch result will be replaced with a raising call
-        async def crashing_run(*args, **kwargs):
-            raise RuntimeError("kaboom: database connection lost")
-
-        executor.dispatch_food_truck = crashing_run  # type: ignore[method-assign]
-        tool_ctx.executor = executor
-
-        from autoskillit.fleet._api import execute_dispatch
-        from tests.fleet._helpers import (
-            _no_sleep_quota_checker,
-            _noop_quota_refresher,
-            _simple_prompt_builder,
-        )
-
-        result = await execute_dispatch(
-            tool_ctx=tool_ctx,
-            recipe="test-recipe",
-            task="do something",
-            ingredients=None,
-            dispatch_name="dispatch-a",
-            timeout_sec=None,
-            prompt_builder=_simple_prompt_builder,
-            quota_checker=_no_sleep_quota_checker,
-            quota_refresher=_noop_quota_refresher,
-        )
-        from autoskillit.server.tools.tools_fleet_dispatch import _write_dispatch_to_campaign_state
-
-        _write_dispatch_to_campaign_state(
-            str(campaign_state_path), "dispatch-a", result.outcome, result.per_dispatch_state_path
-        )
-
-        envelope = json.loads(result.outcome.to_envelope())
-        assert envelope["success"] is False
-        assert envelope["error"] == "fleet_l3_startup_or_crash"
-        assert "RuntimeError" in envelope["user_visible_message"]
-        assert "kaboom: database connection lost" in envelope["user_visible_message"]
-
-        state = read_state(campaign_state_path)
-        assert state is not None
-        refused = [d for d in state.dispatches if d.status.value == "refused"]
-        assert len(refused) == 1
-        assert "RuntimeError" in refused[0].diagnostic_message
-        assert "kaboom: database connection lost" in refused[0].diagnostic_message
-        assert refused[0].reason == "fleet_l3_startup_or_crash"
-
-    async def test_crash_logs_structured_fields(self, tool_ctx, monkeypatch) -> None:
-        """WARNING log must include exc_type and dispatch_name."""
-        import structlog
-
-        from tests.fakes import InMemoryHeadlessExecutor
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        executor = InMemoryHeadlessExecutor()
-
-        async def crashing_run(*args, **kwargs):
-            raise TypeError("unexpected None in argument")
-
-        executor.dispatch_food_truck = crashing_run  # type: ignore[method-assign]
-        tool_ctx.executor = executor
-
-        from autoskillit.fleet._api import execute_dispatch
-        from tests.fleet._helpers import (
-            _no_sleep_quota_checker,
-            _noop_quota_refresher,
-            _simple_prompt_builder,
-        )
-
-        with structlog.testing.capture_logs() as cap:
-            result = await execute_dispatch(
-                tool_ctx=tool_ctx,
-                recipe="test-recipe",
-                task="do something",
-                ingredients=None,
-                dispatch_name="my-test-dispatch",
-                timeout_sec=None,
-                prompt_builder=_simple_prompt_builder,
-                quota_checker=_no_sleep_quota_checker,
-                quota_refresher=_noop_quota_refresher,
-            )
-
-        envelope = json.loads(result.outcome.to_envelope())
-        assert envelope["error"] == "fleet_l3_startup_or_crash"
-
-        warning_logs = [
-            r
-            for r in cap
-            if r.get("event", "") == "execute_dispatch crashed before dispatch completion"
-        ]
-        assert len(warning_logs) >= 1
-        log_record = warning_logs[0]
-        assert log_record.get("exc_type") == "TypeError"
-        assert log_record.get("dispatch_name") == "my-test-dispatch"
-
-    async def test_reject_with_state_persists_message_to_both_states(
-        self, tool_ctx, monkeypatch, tmp_path: Path
-    ) -> None:
-        """_reject_with_state persists diagnostic message to both state files.
-
-        The diagnostic message should be written to both the per-dispatch state file
-        and the campaign state file when a dispatch is rejected.
-        """
-        from autoskillit.fleet import (
-            DispatchRecord,
-            read_state,
-            write_initial_state,
-        )
-
-        campaign_state_path = tmp_path / "campaign.json"
-        write_initial_state(
-            campaign_state_path,
-            "cmp-reject",
-            "test-campaign",
-            "/m.yaml",
-            [DispatchRecord(name="dispatch-b")],
-        )
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        tool_ctx.kitchen_id = "cmp-reject"
-
-        from autoskillit.fleet._api import execute_dispatch
-        from tests.fleet._helpers import (
-            _no_sleep_quota_checker,
-            _noop_quota_refresher,
-            _simple_prompt_builder,
-        )
-
-        result = await execute_dispatch(
-            tool_ctx=tool_ctx,
-            recipe="test-recipe",
-            task="do something",
-            ingredients={"unknown_ingredient_key": "value"},
-            dispatch_name="dispatch-b",
-            timeout_sec=None,
-            prompt_builder=_simple_prompt_builder,
-            quota_checker=_no_sleep_quota_checker,
-            quota_refresher=_noop_quota_refresher,
-        )
-        from autoskillit.server.tools.tools_fleet_dispatch import _write_dispatch_to_campaign_state
-
-        _write_dispatch_to_campaign_state(
-            str(campaign_state_path), "dispatch-b", result.outcome, result.per_dispatch_state_path
-        )
-
-        envelope = json.loads(result.outcome.to_envelope())
-        assert envelope["success"] is False
-        assert "Unknown ingredient keys" in envelope["user_visible_message"]
-
-        campaign_state = read_state(campaign_state_path)
-        assert campaign_state is not None
-        refused_campaign = [d for d in campaign_state.dispatches if d.status.value == "refused"]
-        assert len(refused_campaign) == 1
-        assert "Unknown ingredient keys" in refused_campaign[0].diagnostic_message
-
-
-class TestLabelsCleanedFieldPersistence:
-    @pytest.mark.anyio
-    async def test_labels_cleaned_true_when_cleanup_succeeds(
-        self, tool_ctx, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """DispatchRecord persists labels_cleaned=True only when cleanup actually succeeds."""
-        import dataclasses
-
-        from autoskillit.fleet.sidecar import sidecar_path as make_sidecar_path
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        swap_labels_mock = AsyncMock(return_value={"success": True})
-        github_client = AsyncMock()
-        github_client.swap_labels = swap_labels_mock
-        tool_ctx.github_client = github_client
-
-        failure_result = dataclasses.replace(
-            _DEFAULT_SKILL_RESULT,
-            success=False,
-            result='{"success": false, "reason": "context_exhaustion"}',
-            subtype="success",
-            is_error=False,
-            exit_code=0,
-        )
-
-        async def _return_failure(**kwargs):
-            sidecar = make_sidecar_path(kwargs["dispatch_id"], tool_ctx.project_dir)
-            sidecar.write_text(
-                json.dumps(
-                    {
-                        "issue_url": "https://github.com/owner/repo/issues/1",
-                        "status": "completed",
-                        "ts": "2026-01-01T00:00:00Z",
-                    }
-                )
-                + "\n"
-            )
-            if kwargs.get("on_spawn"):
-                kwargs["on_spawn"](12345, 1000)
-            return failure_result
-
-        tool_ctx.executor.dispatch_food_truck = _return_failure
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_no_sentinel(),
-        )
-
-        await _run(tool_ctx)
-
-        record = _read_dispatch_record(tool_ctx)
-        assert record["status"] == "failure"
-        assert record["labels_cleaned"] is True
-
-    @pytest.mark.anyio
-    async def test_labels_cleaned_false_when_cleanup_fails(
-        self, tool_ctx, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """DispatchRecord persists labels_cleaned=False when cleanup fails."""
-        import dataclasses
-
-        from autoskillit.fleet.sidecar import sidecar_path as make_sidecar_path
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        swap_labels_mock = AsyncMock(side_effect=Exception("rate limited"))
-        github_client = AsyncMock()
-        github_client.swap_labels = swap_labels_mock
-        tool_ctx.github_client = github_client
-
-        failure_result = dataclasses.replace(
-            _DEFAULT_SKILL_RESULT,
-            success=False,
-            result='{"success": false, "reason": "context_exhaustion"}',
-            subtype="success",
-            is_error=False,
-            exit_code=0,
-        )
-
-        async def _return_failure(**kwargs):
-            sidecar = make_sidecar_path(kwargs["dispatch_id"], tool_ctx.project_dir)
-            sidecar.write_text(
-                json.dumps(
-                    {
-                        "issue_url": "https://github.com/owner/repo/issues/1",
-                        "status": "completed",
-                        "ts": "2026-01-01T00:00:00Z",
-                    }
-                )
-                + "\n"
-            )
-            if kwargs.get("on_spawn"):
-                kwargs["on_spawn"](12345, 1000)
-            return failure_result
-
-        tool_ctx.executor.dispatch_food_truck = _return_failure
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_no_sentinel(),
-        )
-
-        await _run(tool_ctx)
-
-        record = _read_dispatch_record(tool_ctx)
-        assert record["status"] == "failure"
-        assert record["labels_cleaned"] is False
-
-    @pytest.mark.anyio
-    async def test_labels_cleaned_false_on_success_outcome(
-        self, tool_ctx, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """DispatchRecord persists labels_cleaned=False when outcome is SUCCESS."""
-        _setup_dispatch(tool_ctx, monkeypatch)
-        tool_ctx.github_client = None
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_completed_clean(True),
-        )
-
-        await _run(tool_ctx)
-
-        record = _read_dispatch_record(tool_ctx)
-        assert record["status"] == "success"
-        assert record["labels_cleaned"] is False
-
-    @pytest.mark.anyio
-    async def test_labels_cleaned_true_with_no_client_no_sidecar(
-        self, tool_ctx, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        """labels_cleaned=True when github_client is None (vacuously true — nothing to clean)."""
-        import dataclasses
-
-        from tests.fakes import _DEFAULT_SKILL_RESULT
-
-        _setup_dispatch(tool_ctx, monkeypatch)
-        tool_ctx.github_client = None
-
-        failure_result = dataclasses.replace(
-            _DEFAULT_SKILL_RESULT,
-            success=False,
-            result='{"success": false, "reason": "context_exhaustion"}',
-            subtype="success",
-            is_error=False,
-            exit_code=0,
-        )
-        tool_ctx.executor = InMemoryHeadlessExecutor(default_result=failure_result)
-        monkeypatch.setattr(
-            "autoskillit.fleet._api.parse_l3_result_block",
-            lambda **_: _make_no_sentinel(),
-        )
-
-        await _run(tool_ctx)
-
-        record = _read_dispatch_record(tool_ctx)
-        assert record["status"] == "failure"
-        assert record["labels_cleaned"] is True
